backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
python-dotenv
pytest
pytest-asyncio
httpx
pandas
openpyxl
google-cloud-translate==3.11.1
google-cloud-texttospeech==2.16.3 